+++ src/zynq/libxdma.c
@@ -311,7 +311,7 @@ xdma_status xdmaFree(xdma_buf_handle handle) {
     return ret;
 }
 
-static inline xdma_status _xdmaStream(xdma_buf_handle buffer, size_t len, unsigned int offset,
+static inline xdma_status _xdmaStream(xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_device dev, xdma_channel ch, bool block, xdma_transfer_handle *transfer)
 {
     struct xdma_chan_cfg *channel = (struct xdma_chan_cfg*)ch;
@@ -365,19 +365,19 @@ static inline xdma_status _xdmaStream(xdma_buf_handle buffer, size_t len, unsign
     return XDMA_SUCCESS;
 }
 
-xdma_status xdmaStream(xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaStream(xdma_buf_handle buffer, size_t len, size_t offset,
             xdma_device dev, xdma_channel channel)
 {
     return _xdmaStream(buffer, len, offset, dev, channel, 1 /*block*/, NULL /*xdma_transfer_handle*/);
 }
 
-xdma_status xdmaStreamAsync(xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaStreamAsync(xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_device dev, xdma_channel channel, xdma_transfer_handle *transfer)
 {
     return _xdmaStream(buffer, len, offset, dev, channel, 0 /*block*/, transfer);
 }
 
-xdma_status xdmaMemcpy(void *usr, xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaMemcpy(void *usr, xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_dir mode)
 {
     alloc_info_t *info = (alloc_info_t *)buffer;
@@ -432,7 +432,7 @@ xdma_status xdmaMemcpy(void *usr, xdma_buf_handle buffer, size_t len, unsigned i
     return ret;
 }
 
-xdma_status xdmaMemcpyAsync(void *usr, xdma_buf_handle buffer, size_t len, unsigned int offset,
+xdma_status xdmaMemcpyAsync(void *usr, xdma_buf_handle buffer, size_t len, size_t offset,
         xdma_dir mode, xdma_transfer_handle *transfer)
 {
     *transfer = (xdma_transfer_handle)NULL;
